requirements.txt
absl-py==0.9.0
aiofiles==0.4.0
appdirs==1.4.4
asgiref==3.2.7
astor==0.8.1
attrs==19.3.0
autobahn==20.4.2
Automat==20.2.0
biopython==1.77
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
click==7.1.1
constantly==15.1.0
cryptography==2.9
daphne==2.5.0
distlib==0.3.0
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dnspython==1.16.0
fastapi==0.52.0
filelock==3.0.12
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.1
h11==0.9.0
h5py==2.10.0
httptools==0.1.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.6.0
incremental==17.5.0
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2
mock==4.0.2
mysql-connector-python==8.0.19
mysqlclient==1.4.6
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.5.3.56
opt-einsum==3.1.0
pandas==1.3.1
Pillow==8.3.1
pipenv==2020.6.2
protobuf==3.6.1
psutil==5.7.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.4
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pytesseract==0.3.8
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
robotframework==3.2.2
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
service-identity==18.1.0
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
starlette==0.13.2
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
torch==1.9.0
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.10.0.0
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
videocr==0.1.6
virtualenv==20.0.20
virtualenv-clone==0.5.4
websockets==8.1
Werkzeug==1.0.0
wrapt==1.11.2
zipp==3.1.0
zope.interface==5.1.0
